requirements.txt
# Ravens Perch v3 - Dependencies
flask>=2.0
pyudev>=0.24
av>=10.0
pyturbojpeg>=1.7
ruamel.yaml>=0.17
requests>=2.28
psutil>=5.9
